Show the Annotations of a signature object

This funtion prints out the information about the call used to compute the drug signatures, and the session info for the session in which the computation was done. Useful for determining the exact conditions used to generate signatures.

showSigAnnot(Sigs)

Arguments

Sigs

An object of the ToxicoSig Class, as returned by drugPerturbationSig

Value

Prints the ToxicoGx Signatures annotations to the output stream, and returns invisible NULL.

Examples

data(TGGATESsmall)
drug.perturbation <- drugPerturbationSig(TGGATESsmall, mDataType="rna", nthread=1, duration = "2",
     drugs = head(treatmentNames(TGGATESsmall)), features = fNames(TGGATESsmall, "rna")[seq_len(2)])
showSigAnnot(drug.perturbation)
#> [1] "drugPerturbationSig"                      
#> [2] "TGGATESsmall"                             
#> [3] "rna"                                      
#> [4] "head(treatmentNames(TGGATESsmall))"       
#> [5] "fNames(TGGATESsmall, \"rna\")[seq_len(2)]"
#> [6] "2"                                        
#> [7] "1"                                        
#> R version 4.2.1 (2022-06-23)
#> Platform: x86_64-pc-linux-gnu (64-bit)
#> Running under: Ubuntu 22.04.2 LTS
#> 
#> Matrix products: default
#> BLAS:   /usr/lib/x86_64-linux-gnu/openblas-pthread/libblas.so.3
#> LAPACK: /usr/lib/x86_64-linux-gnu/openblas-pthread/libopenblasp-r0.3.20.so
#> 
#> locale:
#>  [1] LC_CTYPE=C.UTF-8       LC_NUMERIC=C           LC_TIME=C.UTF-8       
#>  [4] LC_COLLATE=C           LC_MONETARY=C.UTF-8    LC_MESSAGES=C.UTF-8   
#>  [7] LC_PAPER=C.UTF-8       LC_NAME=C              LC_ADDRESS=C          
#> [10] LC_TELEPHONE=C         LC_MEASUREMENT=C.UTF-8 LC_IDENTIFICATION=C   
#> 
#> attached base packages:
#> [1] stats4    stats     graphics  grDevices utils     datasets  methods  
#> [8] base     
#> 
#> other attached packages:
#>  [1] ggplot2_3.4.1               ToxicoGx_2.3.2             
#>  [3] CoreGx_2.2.0                SummarizedExperiment_1.28.0
#>  [5] Biobase_2.58.0              GenomicRanges_1.50.2       
#>  [7] GenomeInfoDb_1.34.9         IRanges_2.32.0             
#>  [9] S4Vectors_0.36.2            MatrixGenerics_1.10.0      
#> [11] matrixStats_0.63.0          BiocGenerics_0.44.0        
#> 
#> loaded via a namespace (and not attached):
#>   [1] fgsea_1.24.0                colorspace_2.1-0           
#>   [3] ellipsis_0.3.2              rprojroot_2.0.3            
#>   [5] lsa_0.73.3                  XVector_0.38.0             
#>   [7] fs_1.6.1                    rstudioapi_0.14            
#>   [9] SnowballC_0.7.0             MultiAssayExperiment_1.24.0
#>  [11] DT_0.27                     fansi_1.0.4                
#>  [13] xml2_1.3.3                  codetools_0.2-18           
#>  [15] downlit_0.4.2               cachem_1.0.7               
#>  [17] knitr_1.42                  jsonlite_1.8.4             
#>  [19] cluster_2.1.3               shinydashboard_0.7.2       
#>  [21] shiny_1.7.4                 compiler_4.2.1             
#>  [23] httr_1.4.5                  backports_1.4.1            
#>  [25] assertthat_0.2.1            Matrix_1.4-1               
#>  [27] fastmap_1.1.1               bench_1.1.2                
#>  [29] limma_3.54.2                cli_3.6.0                  
#>  [31] later_1.3.0                 visNetwork_2.1.2           
#>  [33] htmltools_0.5.4             tools_4.2.1                
#>  [35] igraph_1.4.1                gtable_0.3.1               
#>  [37] glue_1.6.2                  GenomeInfoDbData_1.2.9     
#>  [39] reshape2_1.4.4              dplyr_1.1.0                
#>  [41] fastmatch_1.1-3             Rcpp_1.0.10                
#>  [43] slam_0.1-50                 pkgdown_2.0.7              
#>  [45] vctrs_0.5.2                 BumpyMatrix_1.6.0          
#>  [47] xfun_0.37                   stringr_1.5.0              
#>  [49] mime_0.12                   lifecycle_1.0.3            
#>  [51] gtools_3.9.4                zlibbioc_1.44.0            
#>  [53] scales_1.2.1                ragg_1.2.5                 
#>  [55] promises_1.2.0.1            relations_0.6-13           
#>  [57] parallel_4.2.1              sets_1.0-24                
#>  [59] yaml_2.3.7                  curl_5.0.0                 
#>  [61] memoise_2.0.1               downloader_0.4             
#>  [63] stringi_1.7.12              desc_1.4.2                 
#>  [65] checkmate_2.1.0             caTools_1.18.2             
#>  [67] BiocParallel_1.32.5         rlang_1.0.6                
#>  [69] pkgconfig_2.0.3             systemfonts_1.0.4          
#>  [71] bitops_1.0-7                evaluate_0.20              
#>  [73] lattice_0.20-45             purrr_1.0.1                
#>  [75] htmlwidgets_1.6.1           cowplot_1.1.1              
#>  [77] tidyselect_1.2.0            plyr_1.8.8                 
#>  [79] magrittr_2.0.3              R6_2.5.1                   
#>  [81] gplots_3.1.3                generics_0.1.3             
#>  [83] DelayedArray_0.24.0         pillar_1.8.1               
#>  [85] whisker_0.4.1               withr_2.5.0                
#>  [87] RCurl_1.98-1.10             tibble_3.2.0               
#>  [89] crayon_1.5.2                KernSmooth_2.23-20         
#>  [91] utf8_1.2.3                  rmarkdown_2.20             
#>  [93] grid_4.2.1                  data.table_1.14.8          
#>  [95] marray_1.76.0               piano_2.14.0               
#>  [97] digest_0.6.31               xtable_1.8-4               
#>  [99] tidyr_1.3.0                 httpuv_1.6.9               
#> [101] textshaping_0.3.6           munsell_0.5.0              
#> [103] shinyjs_2.1.0